Add spec file and stub so devs can generate executables
I decided to use the "single-file" option for pyinstaller, which
generates a single .exe file, with everything else bundled (in
compressed form). This seems like the easiest way to distribute FINESSE
in the short time, but in future we may want to consider something else
in order to get faster start times.

Closes #45.