Properties monitored by EM27 now handled as a dictionary.
Anticipating the data scraped off the webpage being stored in a dictionary, method to poll the server now expects a dictionary, and methods to set the text in each QLineEdit are called using the dictionary keys, which expect the respective value. Placeholder function to retrieve values from server updated to return a dictionary.