Add async version of wait_until_stopped
This is needed so that the GUI can be made aware when a movement is complete. In particular, we need this for running measure scripts, as we only want measuring to commence after the motor has stopped.

Now, the stepper.move.end message is sent whenever a movement is finished.