Use a separate message for notify_on_stopped
When a move is triggered by clicking buttons in the GUI, the
notification is unnecessary. It's really only needed for the measure
scripts and by *only* requesting notifications in this context, we can
ensure that only movements triggered from within the measure script code
will elicit this response, which will simplify the handling.