Improved and extended functionality of DP9800 class
Separated out methods to read data from device and parsing it; added data and sysflag attributes to facilitate this. parse() method extracts sysflag and print_sysflag() provides system info using this. read() method now calculates the BCC and compares this to that sent. Provided additional documentation in method docstring to explain this. Added a subclass to create a test/dummy DP9800 device.