Integrated LEDIcons, minor changes to layout.
Merged in LEDIcon module, so poll icon on EM27 monitor now uses this. Set the size policy of this LEDIcon instance to that of the text boxes on this panel for equal spacing and sizing of all rows. Set the left-hand side of the main window to follow a two-column grid layout.