Improved logic of updating widgets
QLineEdits which display EM27 monitored properties are now stored in a dictionary and can therefore be indexed according to their corresponding identifiers.
Also separated the widgets related to the properties from the poll server label and indicator light, so as to avoid shuffling the poll server widgets to the bottom of the layout on each update.