Fix log spam from `EM27Scraper` and `DummyEM27Scraper`
Currently the `EM27Scraper` logs every time it loads data, which is every other second or so. This means that if you're running the program your console log will very quickly be filled up with these messages obscuring whatever other output you might be looking for.

My bad for not clocking this during the code review.