Improvements to blackbody temperature figure
Plot now shows 15 minutes worth of data, as per Jon's request. This is achieved by adopting a poll interval of 2 seconds and thus setting the number of points that will be shown in 15 minutes.
Also added significant improvements to the scaling of the y-axes. Separated this into a separate method so that it can be re-used: it is called when new data is added but also when an individual plot visibility is toggled.