Use persistent attributes to distinguish USB serial devices
Currently, devices are distinguished based on their COM-port number (or
the Unix equivalent), which is not guaranteed to be constitent between
runs. For example, on Windows the COM-port number is fixed to specific
USB ports for a given PC.

Change the USB serial code to use various persistent attributes
of devices instead: viz. vendor ID, product ID and serial number (which
is sometimes, but not always present). Devices which provide multiple
serial ports (of which we are using one for FINESSE at present) will
yield ports with the same attributes, so an additional "count" field is
also used to distinguish devices in this case. It appears in this latter
case that the serial ports will be in the same order for a given device,
so this should be robust (e.g. if the stepper motor is the second port
for this device, it will be always be given a higher COM-port number
than the first).

Closes #323.