Clean up `Device` inheritance logic
Currently, in order to create a device base type, the user passes
`is_base_type=True` to the class definition. If this parameter is
omitted, it is assumed that the class is a concrete device type, in
which case, the class must inherit *directly* from a base type.

There are a couple of problems with this:

1. It doesn't allow for classes that inherit indirectly from a base type
2. The extra `is_base_type` parameter is a bit ugly

I've reworked the logic a bit to fix this. Now, if a class inherits
directly from `Device` it is treated as a base type (no extra argument
necessary) and otherwise it is treated as a device type *only* if it is
not abstract (abstract classes are ignored).

I have a use case for this: I want to have an `OPUSInterfaceBase` class
which inherits from a `SpectrometerBase` class. This class will be
neither a device type or a base device type, but its children will be
device types.