Fix stepper motor port for FINESSE
I think I must have had the port wrong before, but it's a bit annoying
that I didn't spot it, as now it is a bug in v1.2. I was a little
worried that I *hadn't* made a mistake, but that the supposedly
persistent USB strings I'm using aren't actually persistent. It doesn't
seem like that's the case though: the ports are in the same order on two
different machines (running Windows + Linux).