Drop use of USB serial numbers
Owing to #458, we cannot rely on USB serial numbers on Windows (at least
until the upstream bug is fixed). In the interim, drop the use of serial
numbers to distinguish devices and just use the vendor and product IDs.
I have confirmed with @jonemurray that both the FINESSE and UNIRAS rigs
will use serial adaptors of different makes/models, so these IDs will be
different.

Once the upstream bug is fixed, we should revert this change, otherwise
users will not be able to use two USB serial adaptors of the same make
and model at the same time (or, at least, they won't be able to use the
hardware set functionality).

Closes #458.