Automatically generate state machine diagrams for documentation
Currently there is a diagram for the `ScriptRunner` state machine in the documentation, but it was statically generated and does not update automatically when the code changes. In addition, there is another state machine in the code (`OPUSStateMachine`) which does not have a diagram.

Add a mkdocs plugin to automatically generate these diagrams.