Fix: Move device error message code to HardwareSetsControl
Currently, the code responsible for showing an error message when a device error occurs lives in the DeviceControl class, where it always has. The problem is that nowadays a DeviceControl is only created when the user clicks "Manage devices", meaning that otherwise no error dialogs will be shown for device errors.

Fix by moving the relevant code to the HardwareSetsControl class instead.