Refresh list of USB serial devices if port not found
If the user starts FINESSE with one or more of the USB devices not connected (or powered off) then they may be unable to use it, even if it is connected later. The only remedy is to restart the program. The reason for this is because the list of serial ports is retrieved once when the SerialDevice class is created and the result is cached.

An easy workaround for this is to force-refresh the list of serial ports if a given port isn't found. This is a small QoL improvement for users.

This relates to #335, but doesn't fix it (the list of serial ports in the GUI won't be updated when they are refreshed).