Fix mypy error in test
It seems the latest version of `mypy` has noticed the trickery I was doing with this method and is complaining about it.

The body of the method is never actually executed -- we just need it because the parent class is abstract -- but let's replace it with something that `mypy` is happier.