Add functional test that all hardware set files are valid
While we test the `_load_builtin_hardware_sets()` does what it's supposed to by using mocking, we don't test that the actual YAML files bundled with FINESSE are valid, which is a potential error source (e.g. if a device class name changes and a config file isn't updated, then we won't notice until a user tries to actually use the file).

Let's just ensure that all the built-in hardware set config files load properly.