Fix sorting of port names without terminating numbers
As reported by @dalonsoa, not all USB device port names end in a number on MacOS, in contrast to Windows and Linux (at least as far as I've seen).

This breaks our current code for sorting the ports, which will throw an error if a port name doesn't end in a number. Fix by making the number optional and also including the rest of the string in the sorting operation.