Reduce precision of temperatures written to CSV file
Currently temperatures will be written with the maximum precision possible for the `Decimal` type, which equates to a lot of decimal places. This wasn't such an issue when using the DP9800 as it uses a protocol with plain text numbers to a limited number of decimal places, but now that we are using the Seneca device with a binary encoding this can end up being a lot longer. Let's just cap the precision to 0.01 Kelvin.