Fix: `DummyStepperMotor` returns `None` if moving
As with the `ST10Controller` before #743, the `DummyStepperMotor` returns an angle of `None` if it is in motion. However, this isn't what we want anymore; if you try to record data while using the `DummyStepperMotor` you now get an error when the code tries to use this `None` value.

Instead, return the old (pre-move) angle.